restructure back to single VC for feed map VC
